requirements.txt
python-dotenv
Flask
Flask-Bcrypt
psycopg2-binary
Flask-SQLAlchemy
Flask-DebugToolbar
Flask-WTF
ipython
boto3
